Update to use new Makefile and manifest
