+++ server/manifest.go
@@ -1,0 +1,9 @@
+package main
+
+var manifest = struct {
+	Id      string
+	Version string
+}{
+	Id:      "jitsi",
+	Version: "1.2.0",
+}
