--- webapp/client/index.js
@@ -1,5 +1,0 @@
-import ClientClass from './client.js';
-
-const Client = new ClientClass();
-
-export default Client;
