--- webapp/constants/index.js
@@ -1,5 +1,0 @@
-import Svgs from './svgs.js';
-
-export {
-    Svgs
-};
