--- webapp/constants/svgs.js
@@ -1,6 +1,0 @@
-export default {
-    VIDEO_CAMERA: "<svg width='14px' height='10px' viewBox='0 0 16 12' version='1.1' xmlns='http://www.w3.org/2000/svg' xmlns:xlink='http://www.w3.org/1999/xlink'> <g stroke='none' stroke-width='1' fill='inherit' fill-rule='evenodd'> <g transform='translate(-696.000000, -34.000000)' fill-rule='nonzero' fill='inherit'> <g transform='translate(-1.000000, 0.000000)'> <g transform='translate(687.000000, 22.000000)'> <g transform='translate(10.000000, 12.000000)'> <path d='M15.105,1.447 L12,3 L12,1 C12,0.447 11.553,0 11,0 L1,0 C0.447,0 0,0.447 0,1 L0,11 C0,11.553 0.447,12 1,12 L11,12 C11.553,12 12,11.553 12,11 L12,9 L15.105,10.553 C15.6,10.8 16,10.553 16,10 L16,2 C16,1.447 15.6,1.2 15.105,1.447 Z M12.895,7.211 C12.612,7.07 12.306,7 12,7 L10.5,7 C10.224,7 10,7.224 10,7.5 L10,10 L2,10 L2,2 L10,2 L10,4.5 C10,4.776 10.224,5 10.5,5 L12,5 C12.306,5 12.612,4.93 12.895,4.789 L14,4.236 L14,7.763 L12.895,7.211 Z'></path> </g> </g> </g> </g> </g> </svg>",
-    VIDEO_CAMERA_2: "<svg width='27px' height='14px' viewBox='0 0 19 10' version='1.1' xmlns='http://www.w3.org/2000/svg' xmlns:xlink='http://www.w3.org/1999/xlink'><title>icon video small</title> <desc>Created with Sketch.</desc><defs><linearGradient x1='50%' y1='0%' x2='50%' y2='100%' id='linearGradient-1'><stop stop-color='#FFFFFF' offset='0%'></stop><stop stop-color='#F4F4F4' offset='100%'></stop></linearGradient></defs><g id='v2' stroke='none' stroke-width='1' fill='inherit' fill-rule='evenodd'><g id='zoom-bot-&amp;-drop-down' transform='translate(-443.000000, -250.000000)'><g id='zoom-bot---share-meeting' transform='translate(340.000000, 98.000000)'><g id='Send-Button' transform='translate(83.000000, 136.000000)'><g id='icon-video-small' transform='translate(20.000000, 16.000000)'><path d='M1,0 L10,0 L10,0 C12.209139,-4.05812251e-16 14,1.790861 14,4 L14,9 L14,9 C14,9.55228475 13.5522847,10 13,10 L4,10 L4,10 C1.790861,10 2.705415e-16,8.209139 0,6 L0,1 L0,1 C-6.76353751e-17,0.44771525 0.44771525,1.01453063e-16 1,0 Z' id='Rectangle-4' fill='inherit'></path><path d='M15.3667766,2.86236308 L17.3667766,1.22606168 L17.3667766,1.22606168 C17.794228,0.876342075 18.4242495,0.93935609 18.7739691,1.36680742 C18.9201398,1.54546732 19,1.76919482 19,2.00003079 L19,8.99999935 L19,8.99999935 C19,9.5522841 18.5522847,9.99999935 18,9.99999935 C17.769164,9.99999935 17.5454365,9.92013917 17.3667766,9.77396845 L15.3667766,8.13766705 L15.3667766,8.13766705 C15.1346326,7.94773826 15,7.66363761 15,7.36369795 L15,3.63633218 L15,3.63633218 C15,3.33639253 15.1346326,3.05229188 15.3667766,2.86236308 Z' id='Rectangle-5' fill='inherit'></path></g></g></g></g></g></svg>",
-    VIDEO_CAMERA_3: "<svg width='19px' height='10px' viewBox='0 0 19 10' version='1.1' xmlns='http://www.w3.org/2000/svg' xmlns:xlink='http://www.w3.org/1999/xlink'><title>icon video small</title><desc>Created with Sketch.</desc><defs><linearGradient x1='50%' y1='0%' x2='50%' y2='100%' id='linearGradient-1'><stop stop-color='#FFFFFF' offset='0%'></stop><stop stop-color='#F4F4F4' offset='100%'></stop></linearGradient></defs><g id='v2' stroke='none' stroke-width='1' fill='inherit' fill-rule='evenodd'><g id='zoom-bot-&amp;-drop-down' transform='translate(-443.000000, -250.000000)'><g id='zoom-bot---share-meeting' transform='translate(340.000000, 98.000000)'><g id='Send-Button' transform='translate(83.000000, 136.000000)'><g id='icon-video-small' transform='translate(20.000000, 16.000000)'><path d='M1,0 L10,0 L10,0 C12.209139,-4.05812251e-16 14,1.790861 14,4 L14,9 L14,9 C14,9.55228475 13.5522847,10 13,10 L4,10 L4,10 C1.790861,10 2.705415e-16,8.209139 0,6 L0,1 L0,1 C-6.76353751e-17,0.44771525 0.44771525,1.01453063e-16 1,0 Z' id='Rectangle-4' fill='inherit'></path><path d='M15.3667766,2.86236308 L17.3667766,1.22606168 L17.3667766,1.22606168 C17.794228,0.876342075 18.4242495,0.93935609 18.7739691,1.36680742 C18.9201398,1.54546732 19,1.76919482 19,2.00003079 L19,8.99999935 L19,8.99999935 C19,9.5522841 18.5522847,9.99999935 18,9.99999935 C17.769164,9.99999935 17.5454365,9.92013917 17.3667766,9.77396845 L15.3667766,8.13766705 L15.3667766,8.13766705 C15.1346326,7.94773826 15,7.66363761 15,7.36369795 L15,3.63633218 L15,3.63633218 C15,3.33639253 15.1346326,3.05229188 15.3667766,2.86236308 Z' id='Rectangle-5' fill='inherit'></path></g></g></g></g></g></svg>",
-    SHARE: "<svg width='27px' height='18px' viewBox='0 0 27 18' version='1.1' xmlns='http://www.w3.org/2000/svg' xmlns:xlink='http://www.w3.org/1999/xlink'><title>icon share</title><desc>Created with Sketch.</desc><defs></defs><g id='v2' stroke='none' stroke-width='1' fill='inherit' fill-rule='evenodd'><g id='zoom-bot-&amp;-drop-down' transform='translate(-807.000000, -122.000000)' fill='inherit'><g id='Pop-up' transform='translate(787.000000, 50.000000)'><path d='M23,72 L44,72 L44,72 C45.6568542,72 47,73.3431458 47,75 L47,87 L47,87 C47,88.6568542 45.6568542,90 44,90 L23,90 C21.3431458,90 20,88.6568542 20,87 L20,75 L20,75 C20,73.3431458 21.3431458,72 23,72 L23,72 Z M35.8749836,82.2499851 C35.4322762,82.2499851 35.0416522,82.3867035 34.7031114,82.6601404 L32.7109287,81.4101434 C32.763012,81.1367065 32.763012,80.8632697 32.7109287,80.5898328 L34.7031114,79.3398358 C35.0416522,79.6132726 35.4322762,79.7499911 35.8749836,79.7499911 C36.3958158,79.7499911 36.8385229,79.5676999 37.2031054,79.2031174 C37.567688,78.8385348 37.7499791,78.3958277 37.7499791,77.8749955 C37.7499791,77.3541633 37.567688,76.9114563 37.2031054,76.5468737 C36.8385229,76.1822911 36.3958158,76 35.8749836,76 C35.3541514,76 34.9114443,76.1822911 34.5468618,76.5468737 C34.1822792,76.9114563 33.9999881,77.3541633 33.9999881,77.8749955 C33.9999881,78.0182243 34.013009,78.1549427 34.0390505,78.2851508 L32.0468677,79.5351478 C31.708327,79.261711 31.3177029,79.1249925 30.8749955,79.1249925 C30.3541633,79.1249925 29.9114563,79.3072837 29.5468737,79.6718662 C29.1822911,80.0364488 29,80.4791559 29,80.9999881 C29,81.5208203 29.1822911,81.9635273 29.5468737,82.3281099 C29.9114563,82.6926925 30.3541633,82.8749836 30.8749955,82.8749836 C31.3177029,82.8749836 31.708327,82.7382652 32.0468677,82.4648283 L34.0390505,83.7148254 C34.013009,83.8450335 33.9999881,83.9817519 33.9999881,84.1249806 C33.9999881,84.6458128 34.1822792,85.0885199 34.5468618,85.4531025 C34.9114443,85.817685 35.3541514,85.9999762 35.8749836,85.9999762 C36.3958158,85.9999762 36.8385229,85.817685 37.2031054,85.4531025 C37.567688,85.0885199 37.7499791,84.6458128 37.7499791,84.1249806 C37.7499791,83.6041484 37.567688,83.1614414 37.2031054,82.7968588 C36.8385229,82.4322762 36.3958158,82.2499851 35.8749836,82.2499851 Z' id='icon-share'></path></g></g></g></svg>"
-};
