+++ webapp/package.json
@@ -2,30 +2,36 @@
   "name": "jitsi",
   "version": "1.1.0",
   "description": "Jitsi audio and video conferencing plugin for Mattermost",
-  "dependencies": {
-    "mattermost-redux": "1.0.1",
-    "prop-types": "15.5.10",
-    "react": "^16.4.1",
-    "react-redux": "^5.0.7",
-    "redux": "^4.0.0",
-    "superagent": "3.5.2"
-  },
-  "devDependencies": {
-    "babel-core": "6.25.0",
-    "babel-loader": "7.1.1",
-    "babel-plugin-transform-runtime": "6.23.0",
-    "babel-polyfill": "6.23.0",
-    "babel-preset-es2015": "6.24.1",
-    "babel-preset-react": "6.24.1",
-    "babel-preset-stage-0": "6.24.1",
-    "css-loader": "^0.28.7",
-    "eslint": "3.17.1",
-    "eslint-plugin-react": "6.10.0",
-    "style-loader": "^0.19.0",
-    "webpack": "2.2.1"
-  },
+  "main": "src/index.jsx",
   "scripts": {
-    "build": "webpack --display-error-details --verbose --progress",
-    "check": "eslint --ext \".jsx\" --ext \".js\" --ignore-pattern node_modules --quiet ."
+    "build": "webpack --mode=production",
+    "lint": "eslint --ignore-pattern node_modules --ignore-pattern dist --ext .js --ext .jsx . --quiet",
+    "fix": "eslint --ignore-pattern node_modules --ignore-pattern dist --ext .js --ext .jsx . --quiet --fix"
+  },
+  "author": "",
+  "license": "",
+  "devDependencies": {
+    "babel-core": "^6.26.3",
+    "babel-eslint": "^8.2.6",
+    "babel-loader": "^7.1.5",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-react": "^6.24.1",
+    "eslint": "^5.2.0",
+    "eslint-import-resolver-webpack": "^0.10.1",
+    "eslint-plugin-import": "^2.13.0",
+    "eslint-plugin-react": "^7.10.0",
+    "file-loader": "3.0.1",
+    "webpack": "^4.16.1",
+    "webpack-cli": "^3.0.8"
+  },
+  "dependencies": {
+    "mattermost-redux": "^1.2.0",
+    "react": "16.4.1",
+    "react-intl": "2.8.0",
+    "react-redux": "5.0.7",
+    "redux": "4.0.0",
+    "superagent": "^5.0.5"
   }
 }
