+++ webapp/src/client/index.js
@@ -1,0 +1,5 @@
+import ClientClass from './client.js';
+
+const Client = new ClientClass();
+
+export default Client;
