+++ webapp/src/constants/index.js
@@ -1,0 +1,5 @@
+import Svgs from './svgs.js';
+
+export {
+    Svgs
+};
