+++ webapp/src/manifest.js
@@ -1,0 +1,2 @@
+export const id = 'jitsi';
+export const version = '1.2.0';
