+++ webapp/webpack.config.js
@@ -1,39 +1,44 @@
+
+var path = require('path');
+
 module.exports = {
-    entry: './index.js',
-    output: {
-        filename: 'dist/jitsi_bundle.js'
+    entry: [
+        './src/index.jsx'
+    ],
+    resolve: {
+        modules: [
+            'src',
+            'node_modules',
+            path.resolve(__dirname)
+        ],
+        extensions: ['*', '.js', '.jsx']
     },
     module: {
-        loaders: [
+        rules: [
             {
-                test: /\.(js|jsx)?$/,
-                loader: 'babel-loader',
-                exclude: /(node_modules|non_npm_dependencies)/,
-                query: {
-                    presets: [
-                        'react',
-                        ['es2015', {modules: false}],
-                        'stage-0'
-                    ],
-                    plugins: ['transform-runtime']
-                }
-            },
-            {
-                test: /\.css$/,
-                use: [
-                    {
-                        loader: 'style-loader'
-                    },
-                    {
-                        loader: 'css-loader'
+                test: /\.(js|jsx)$/,
+                exclude: /node_modules/,
+                use: {
+                    loader: 'babel-loader',
+                    options: {
+                        presets: ['env', 'react'],
+                        plugins: [
+                            'transform-class-properties',
+                            'transform-object-rest-spread'
+                        ]
                     }
-                ]
+                }
             }
         ]
     },
     externals: {
         react: 'React',
         redux: 'Redux',
-        'react-redux': 'ReactRedux',
+        'react-redux': 'ReactRedux'
+    },
+    output: {
+        path: path.join(__dirname, '/dist'),
+        publicPath: '/',
+        filename: 'main.js'
     }
 };
